requirements.txt
beautifulsoup4
feedparser
google-genai
jinja2
python-dotenv
requests
